Refactor statement node and link creation into seperate helpers
